bootkube.sh: add cluster-etcd-operator support
This PR adds support for cluster-etcd-operator to be conditionally disabled. Once enabled
and part of CVO operator payload we can also deploy the operator in an Unmanaged state.
This allows for the operator to make through CI tests without enabling it directly. Once
the operator is deemed stable we can move to Managed state where it will remain.

Basically parts of this logic in this PR are required to allow us a safe way to revert.

Signed-off-by: Sam Batschelet <sbatsche@redhat.com>